backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart

langchain
langchain-community
langchain-core
langchain-groq

pinecone

langchain-google-genai

pyPDF

python-dotenv

requests
pydantic
tqdm

loguru
